move connection to helper class
