+++ src/main/kotlin/io/ionic/libs/ionfiletransferlib/IONFLTRController.kt
@@ -1,5 +1,6 @@
 package io.ionic.libs.ionfiletransferlib
 
+import io.ionic.libs.ionfiletransferlib.helpers.IONFLTRConnectionHelper
 import io.ionic.libs.ionfiletransferlib.helpers.IONFLTRFileHelper
 import io.ionic.libs.ionfiletransferlib.helpers.IONFLTRInputsValidator
 import io.ionic.libs.ionfiletransferlib.helpers.runCatchingIONFLTRExceptions
@@ -7,22 +8,19 @@ import io.ionic.libs.ionfiletransferlib.model.IONFLTRDownloadOptions
 import io.ionic.libs.ionfiletransferlib.model.IONFLTRException
 import io.ionic.libs.ionfiletransferlib.model.IONFLTRProgressStatus
 import io.ionic.libs.ionfiletransferlib.model.IONFLTRTransferComplete
-import io.ionic.libs.ionfiletransferlib.model.IONFLTRTransferHttpOptions
 import io.ionic.libs.ionfiletransferlib.model.IONFLTRTransferResult
 import io.ionic.libs.ionfiletransferlib.model.IONFLTRUploadOptions
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 import kotlinx.coroutines.flow.flowOn
+import kotlinx.coroutines.withContext
 import java.io.BufferedInputStream
 import java.io.BufferedOutputStream
 import java.io.File
 import java.io.FileInputStream
 import java.io.FileOutputStream
-import java.net.HttpURLConnection
-import java.net.URL
 import java.nio.charset.StandardCharsets
-import kotlinx.coroutines.withContext
 
 /**
  * Entry point in IONFileTransferLib-Android
@@ -31,11 +29,13 @@ import kotlinx.coroutines.withContext
  */
 class IONFLTRController internal constructor(
     private val inputsValidator: IONFLTRInputsValidator,
-    private val fileHelper: IONFLTRFileHelper
+    private val fileHelper: IONFLTRFileHelper,
+    private val connectionHelper: IONFLTRConnectionHelper
 ) {
     constructor() : this(
         inputsValidator = IONFLTRInputsValidator(),
-        fileHelper = IONFLTRFileHelper()
+        fileHelper = IONFLTRFileHelper(),
+        connectionHelper = IONFLTRConnectionHelper()
     )
 
     companion object {
@@ -101,7 +101,7 @@ class IONFLTRController internal constructor(
             fileHelper.createParentDirectories(targetFile)
 
             // Setup connection
-            val connection = setupConnection(options.url, options.httpOptions)
+            val connection = connectionHelper.setupConnection(options.url, options.httpOptions)
             
             try {
                 connection.connect()
@@ -183,7 +183,7 @@ class IONFLTRController internal constructor(
             }
             
             // Setup connection
-            val connection = setupConnection(options.url, options.httpOptions)
+            val connection = connectionHelper.setupConnection(options.url, options.httpOptions)
             
             try {
                 val fileSize = file.length()
@@ -349,65 +349,6 @@ class IONFLTRController internal constructor(
             }
         }.getOrThrow()
     }.flowOn(Dispatchers.IO)
-
-    /**
-     * Sets up the HTTP connection with the provided options.
-     */
-    private fun setupConnection(urlString: String, httpOptions: IONFLTRTransferHttpOptions): HttpURLConnection {
-        val url = URL(urlString)
-        val connection = url.openConnection() as HttpURLConnection
-        
-        // Set method
-        connection.requestMethod = httpOptions.method
-        
-        // Set timeouts
-        connection.connectTimeout = httpOptions.connectTimeout
-        connection.readTimeout = httpOptions.readTimeout
-        
-        // Set headers
-        httpOptions.headers.forEach { (key, value) ->
-            connection.setRequestProperty(key, value)
-        }
-        
-        // Set parameters
-        if (httpOptions.params.isNotEmpty()) {
-            val paramString = buildString {
-                httpOptions.params.forEach { (key, values) ->
-                    values.forEach { value ->
-                        if (isNotEmpty()) append("&")
-                        val encodedKey = if (httpOptions.shouldEncodeUrlParams) {
-                            java.net.URLEncoder.encode(key, StandardCharsets.UTF_8.name())
-                        } else key
-                        val encodedValue = if (httpOptions.shouldEncodeUrlParams) {
-                            java.net.URLEncoder.encode(value, StandardCharsets.UTF_8.name())
-                        } else value
-                        append("$encodedKey=$encodedValue")
-                    }
-                }
-            }
-            
-            if (httpOptions.method.equals("GET", ignoreCase = true)) {
-                val separator = if (urlString.contains("?")) "&" else "?"
-                val newUrl = URL("$urlString$separator$paramString")
-                return newUrl.openConnection() as HttpURLConnection
-            } else {
-                connection.doOutput = true
-                connection.outputStream.use { os ->
-                    os.write(paramString.toByteArray())
-                }
-            }
-        }
-        
-        // Set redirect handling
-        connection.instanceFollowRedirects = !httpOptions.disableRedirects
-        
-        // Set SSL factory if provided
-        if (httpOptions.sslSocketFactory != null && connection is javax.net.ssl.HttpsURLConnection) {
-            connection.sslSocketFactory = httpOptions.sslSocketFactory
-        }
-        
-        return connection
-    }
 
     /**
      * Validates the URL and file path for transfer operations.
